(December 2012) Updated experiments, simulators, and calculators.  The html files have the proper headers and footers, and we purged some obsolete files.
TO DO!!!

We need to reconsider these tools.  Every distribution can have a simple calculators, but simulators and experiments are not as well defined. Experiments are simulations, just showing some additional information. Conversely, every simulator could be an experiment by showing the method of
simulation (random quantile, rejection method, simulation in terms of simpler distributions, etc.)

Some distributions have several natural simulators, others only one or two. The binomial can be illustrated in terms of coins, a Galton board, sampling with replacement, a Bernoulli trials "timeline", etc. For the normal distribution, the standard method simulation is via polar coordinates, with an exponentially distributed radius and a uniformly distributed angle.  The simulation could show this graphically.

A suggestion is to reorganize all 3 types of tools, at the top level, not by type (since again, the types don't make sense to me) but rather by distribution. For each distribution, we would have a calculator and at least one (often several) simulators.  In the Navigator, each distribution could be linked to the set of tools for that distribution.

We need to enhance the calculators to display the mean, standard deviation, skewness and kurtosis for each distribution, according to the distributome.xml DB.